+++ CHANGES.md
@@ -3,5 +3,6 @@
 ## v0.5.1
 
 - JSON: Support newlines and indentation in strings
+- JSON: Resolve multiple placeholders in single property
 - Fix: Remove dots from file names if they are generated from another property
 - Fix: Don't lower case file names